+++ src/environments/__init__.py
@@ -2,6 +2,13 @@
 
 import observation_builders
 
+from .continuous_action_mapf import ContinuousActionMAPFEnv
+from .point_mass import PointMass2DEnv
 from .procedural_pathfinding import ProceduralPathfindingEnv
 
-__all__ = ["ProceduralPathfindingEnv", "observation_builders"]
+__all__ = [
+    "ContinuousActionMAPFEnv",
+    "PointMass2DEnv",
+    "ProceduralPathfindingEnv",
+    "observation_builders",
+]